requirements.txt
python-telegram-bot
psutil
requests